+++ book/book-examples/tests/view-example-enum.rs
@@ -4,8 +4,7 @@ use bonsaidb::{
         document::{BorrowedDocument, Document},
         schema::{
             view::{map::ViewMappedValue, EnumKey},
-            Collection, DefaultViewSerialization, Name, ReduceResult, View, ViewMapResult,
-            ViewSchema,
+            Collection, ReduceResult, View, ViewMapResult, ViewSchema,
         },
         Error,
     },
@@ -38,19 +37,11 @@ pub struct BlogPost {
 }
 // ANCHOR_END: struct
 
-#[derive(Debug, Clone)]
-pub struct BlogPostsByCategory;
-
 // ANCHOR: view
-impl View for BlogPostsByCategory {
-    type Collection = BlogPost;
-    type Key = Option<Category>;
-    type Value = u32;
 
-    fn name(&self) -> Name {
-        Name::new("by-category")
-    }
-}
+#[derive(Debug, Clone, View)]
+#[view(collection = BlogPost, key = Option<Category>, value = u32, name = "by-category")]
+pub struct BlogPostsByCategory;
 
 impl ViewSchema for BlogPostsByCategory {
     type View = Self;
@@ -69,8 +60,6 @@ impl ViewSchema for BlogPostsByCategory {
     }
 }
 // ANCHOR_END: view
-
-impl DefaultViewSerialization for BlogPostsByCategory {}
 
 #[allow(unused_variables)]
 #[tokio::test]